ENH Don't add default filterclass if explicitly removed. (#3127)
It can be advantageous, e.g. with archived admin, to have a pre-defined
version filter applied to site tree searching. We can't do that if the
default filter class is always applied.

We can signal that we want to use an explicit pre-defined version filter
by removing the ability to search by other version filters.